Adding my name to this list of signatures
Thank you to the organizers of this petition.  I am deeply concerned about the psychological health and safety of both the children and the parents who were unnecessarily separated while seeking asylum, and I want for all of them to be reunited immediately as the first step in resolving this problem for which I can't find the appropriate words to describe how horrible it is.

Thank you also to everyone who has signed this petition so far, and thank you in advance to everyone who signs in the future.