Add my name
I'm generally fairly cautious with which companies I deal with, so just about every boycott in the last 10 years that means something, I've been left with nothing to boycott.  In this case, I was using github before the M$ influence, and it's snuck up on me.  But I finally get to take a stand on something important.  It would be nice to take a stand vs a bunch of other issues (eg: Australia's immigration/refugee response almost makes the USA's look humane) but this is the stand I can take, that has meaning beyond signing petitions that get ignored.
Also, you can set up gitlab sync'ed projects (confusing that repos are called projects, esp when you use bitbucket) with github repos, and keep them auto synced until the day of action.  And it shows in github, so it's a sign that you didn't just sign, you're prepared to take that action.